Removed helper testing function
